Boost Tool
🎃Discord Boost Tool
Hey There! This is the owner of this Boost Tool. Many Were people were skidding the tool and injecting stealers in this tool. This is the Legitimate Version of this tool. It Doesnot have any kind of stealers stored in it. This Tool is safe and free to use. Donot buy this tool From Anyone it is a open source 
---------
Features

-> fast boosting
-> stock editor
-> cleaned
----------------
Requirements

-> Capmonster key
-> nitro tokens

star and follow <3